+++ app/src/main/java/com/ats/hreasy/model/MyLeaveData.java
@@ -11,7 +11,7 @@ public class MyLeaveData {
     private String leaveDuration;
     private String leaveFromdt;
     private String leaveTodt;
-    private Integer leaveNumDays;
+    private float leaveNumDays;
     private String leaveEmpReason;
     private Integer finalStatus;
     private String circulatedTo;
@@ -91,11 +91,11 @@ public class MyLeaveData {
         this.leaveTodt = leaveTodt;
     }
 
-    public Integer getLeaveNumDays() {
+    public float getLeaveNumDays() {
         return leaveNumDays;
     }
 
-    public void setLeaveNumDays(Integer leaveNumDays) {
+    public void setLeaveNumDays(float leaveNumDays) {
         this.leaveNumDays = leaveNumDays;
     }
